Synchronization is in
This is really simple. The app now just adds an execute-at that is
based on a flat 30ms transmission time per device. It sends that
time along to each workerbee who sync with their internal clock
and execute x milliseconds from when the command is received.

This will probably work fine. We may need to adjust that
transmission constant a bit, but we'll be close I think.