Heavily optimized things
- Moved the mapping of four character modes to mode numbers into
the iOS app. This lets us ditch the large map on the feather side

- Moved to a binary protocol instead of sending string. This lets
us decrease the MTU on the feather side. It also allows us to
greatly decrease the buffer size as well as eliminate the parsing
logic altogether. This is a big overall savings in terms of
memory usage

- Moved the bluetooth library to low  bandwidth mode. Thanks to the
savings above we can use a bit less memory here.